+++ src/routes/profile/+page.svelte
@@ -69,8 +69,6 @@
 					</tr>
 				</tbody>
 			</table>
-
-			<a class="btn btn-primary mt-3" href="/profile/token" download>Download token</a>
 		</div>
 	</div>
 </div>