+++ src/routes/profile/token/+server.js
@@ -1,10 +1,10 @@
 export async function GET({ cookies }) {
 	const cookie = cookies.get('fastapiusersauth');
 
-	return new Response(cookie, {
+	return new Response(JSON.stringify({ token: cookie }), {
 		status: 200,
 		headers: {
-			'Content-Disposition': 'Attachment;filename=fractal-token.txt'
+			'Content-Type': 'application/json'
 		}
 	});
 }