+++ tests/get_token_test.spec.js
@@ -1,0 +1,10 @@
+import { expect, test } from '@playwright/test';
+import { waitPageLoading } from './utils.js';
+
+test('Copy token test', async ({ page }) => {
+	await page.goto('/profile');
+	await waitPageLoading(page);
+	await page.getByRole('button', { name: 'admin@fractal.xy' }).click();
+	await page.getByRole('button', { name: 'Get token' }).click();
+	await expect(page.getByText('Token copied to clipboard')).toBeVisible();
+});